Day 83 of 100: Updating logs
Since this does not happen via GitBash, I am manually uploading updates to the Logs page.

I am now done with the General Adverserial Network portion of my Nanodegree. Only one required topic is left which is Unsupervised learning. For now I have added the reviews for the GAN network as there are a lot of resources in it that can be used for a deeper dive into and also there are some tips on how to set up the hyperparameters.

Will probably update the logs by Day 90 of 100.